Major refactor - Single IIFE file to ESM and Class
 - Rewrite large single IIFE-based class, to newer JS class syntax
 - Split up class into multiple files (constants, utils, etc.)
 - Improve type-definitions and type-safety
 - Improve documentation (comments)
 - Move files out of root and into src
 - Start using TypeScript, emit declaration file (still WIP)
