+++ package.json
@@ -29,7 +29,8 @@
   "devDependencies": {
     "@types/jest": "^26.0.23",
     "jest": "^27.0.4",
-    "parcel-bundler": "^1.12.4"
+    "parcel-bundler": "^1.12.4",
+    "typescript": "^4.3.2"
   },
   "dependencies": {},
   "resolutions": {
